로그인 팝업 위치 auth 폴더 하위로 위치 변경 및 로그인/회원가입 step 추가
